+++ cobrafuzz/fuzzer.py
@@ -221,7 +221,6 @@ class Fuzzer:
         state_file: Optional[Path] = None,
         load_crashes: bool = True,
         simplify: Optional[Path] = None,
-        simp_steps: int = 10000,
     ):
         """
         Fuzz-test target and store crash artifacts into crash_dir.
@@ -242,14 +241,13 @@ class Fuzzer:
         max_insert_length:  Maximum number of bytes to insert during mutation.
         max_modifications:  Maximum number of consecutive modifications during mutation.
         max_runs:           Number of target executions after which to exit the fuzzer.
-        max_time:           Number of seconds after which to exit the fuzzer.
+        max_time:           Number of seconds after which to exit.
         num_workers:        Number of parallel workers.
                             Use None to spawn one fewer than CPUs available.
         regression:         Execute target on all samples found in crash_dir, print errors and exit.
         seeds:              List of files and directories to seed the fuzzer with.
         simplify:           When set, run simplifier and store simplified crash results in
                             directory.
-        simp_steps:         Number of unsuccessful steps before stopping simplification.
         start_method:       Multiprocessing start method to use (spawn, forkserver or fork).
                             Defaults to "spawn". Do not use "fork" as it is unreliable and may lead
                             to deadlocks.
@@ -295,7 +293,6 @@ class Fuzzer:
             file=state_file,
         )
         self._simplify = simplify
-        self._simp_steps = simp_steps
 
         if load_crashes:
             self._load_crashes(regression=regression)
@@ -488,7 +485,7 @@ class Fuzzer:
                     crash_dir=self._crash_dir,
                     target=self._target,
                     output_dir=self._simplify,
-                    steps=self._simp_steps,
+                    max_time=self._max_time,
                 ).simplify()
             sys.exit(1)