+++ tests/unit/test_simplifier.py
@@ -1,14 +1,23 @@
 # ruff: noqa: SLF001
 from __future__ import annotations
 
+import copy
 import logging
+import multiprocessing as mp
 from pathlib import Path
-from typing import Callable, Optional
+from typing import Callable, Optional, Tuple, cast
 
+import dill  # type: ignore[import-untyped]
 import pytest
 
 from cobrafuzz import common, simplifier, util
-from tests.utils import StaticRand, do_raise
+from tests import utils
+
+ArgsType = Tuple[
+    bytes,
+    utils.DummyQueue[simplifier.Metrics],
+    utils.DummyQueue[simplifier.Metrics],
+]
 
 
 @pytest.mark.parametrize(
@@ -95,8 +104,8 @@ def test_remove_lines(data: bytes, start: int, end: int, expected: bytes) -> Non
     result = simplifier._simplify_remove_lines(
         data,
         util.Params(
-            start=StaticRand(start),
-            end=StaticRand(end),
+            start=utils.StaticRand(start),
+            end=utils.StaticRand(end),
         ),
     )
     assert result == bytearray(expected)
@@ -237,8 +246,8 @@ def test_remove_characters(data: bytes, start: int, length: int, expected: bytes
     result = simplifier._simplify_remove_characters(
         data,
         util.Params(
-            start=StaticRand(start),
-            length=StaticRand(length),
+            start=utils.StaticRand(start),
+            length=utils.StaticRand(length),
         ),
     )
     assert result == bytearray(expected)
@@ -302,8 +311,8 @@ def test_shorten_token(data: bytes, pos: int, expected: bytes) -> None:
     result = simplifier._simplify_shorten_token(
         data,
         util.Params(
-            pos=StaticRand(pos),
-            pattern=StaticRand(1),
+            pos=utils.StaticRand(pos),
+            pattern=utils.StaticRand(1),
         ),
     )
     assert result == bytearray(expected)
@@ -353,17 +362,16 @@ def test_simplify_invalid_sample(
     (crash_dir / "unrelated").write_bytes(b"unrelated")
     (crash_dir / "invalid").write_bytes(b"invalid")
 
-    s = simplifier.Simp(  # pragma: no cover
-        crash_dir=crash_dir,
-        output_dir=output_dir,
-        target=lambda _: None,
-    )
-
     with monkeypatch.context() as m:
+        s = simplifier.Simp(  # pragma: no cover
+            crash_dir=crash_dir,
+            output_dir=output_dir,
+            target=lambda _: None,
+        )
         m.setattr(
             s,
             "_simplify",
-            lambda d: do_raise(
+            lambda d: utils.do_raise(
                 common.InvalidSampleError,
                 cond=(b"invalid" in d),
                 message="Invalid sample",
@@ -371,6 +379,7 @@ def test_simplify_invalid_sample(
         )
         with caplog.at_level(logging.INFO):
             s.simplify()
+        del s
 
     assert "Could not simplify unrelated" in caplog.text
     assert "Invalid sample: invalid" in caplog.text
@@ -473,7 +482,6 @@ def test_simplify_no_crash(
         crash_dir=tmp_path / "crash",
         output_dir=tmp_path / "output",
         target=target,
-        steps=10,
     )
     with pytest.raises(common.InvalidSampleError, match="^No exception for sample$"):
         s._simplify(data)
@@ -486,9 +494,7 @@ def test_simplify_no_crash(
             b"""START
             CRASH
             END""",
-            b"""START
-            CRASH
-            END""",
+            None,
             lambda data, _: data,
         ),
         (
@@ -505,9 +511,7 @@ def test_simplify_no_crash(
             b"""START
             CRASH
             END""",
-            b"""START
-            CRASH
-            END""",
+            None,
             lambda data, _: b"\n".join(d for d in data.split(b"\n") if b"CRASH" not in d),
         ),
         (
@@ -515,18 +519,13 @@ def test_simplify_no_crash(
             CRASH
             BOOM
             END""",
-            b"""START
-            CRASH
-            BOOM
-            END""",
+            None,
             lambda data, _: b"\n".join(d for d in data.split(b"\n") if b"CRASH" not in d),
         ),
     ],
     ids=range(1, 5),
 )
-def test_simplify(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
+def test_simplification(
     data: bytes,
     expected: Optional[bytes],
     mutator: Callable[[bytes, util.Params], bytes],
@@ -537,13 +536,95 @@ def test_simplify(
             util.Params,
         ]
     ] = util.AdaptiveChoiceBase(population=[(mutator, util.Params())])
+    request_queue: mp.Queue[Optional[simplifier.Metrics]] = mp.Queue()
+    result_queue: mp.Queue[Optional[simplifier.Metrics]] = mp.Queue()
+    baseline = simplifier.run_target(target=target, data=data)
 
-    s = simplifier.Simp(  # pragma: no cover
-        crash_dir=tmp_path / "crash",
-        output_dir=tmp_path / "output",
-        target=target,
-        steps=10,
+    request_queue.put(baseline)
+    request_queue.put(None)
+    simplifier.simplifier(
+        target_bytes=dill.dumps(target),
+        request_queue=request_queue,
+        result_queue=result_queue,
+        mutators=mutators,
     )
-    with monkeypatch.context() as m:
-        m.setattr(s, "_mutators", mutators)
-        assert s._simplify(data) == expected
+
+    result = result_queue.get()
+    if result is None:
+        assert expected is None
+    else:
+        assert result.data == expected
+
+
+def test_terminate_workers(monkeypatch: pytest.MonkeyPatch) -> None:
+    def target(_: bytes) -> None:
+        pass  # pragma: no cover
+
+    args: ArgsType = (
+        b"deadbeef",
+        utils.DummyQueue(),
+        utils.DummyQueue(),
+    )
+
+    workers: list[tuple[utils.DummyProcess[ArgsType], utils.DummyQueue[simplifier.Metrics]]] = [
+        (utils.DummyProcess(target=target, args=args), utils.DummyQueue()),
+        (utils.DummyProcess(target=target, args=args), utils.DummyQueue()),
+    ]
+
+    with monkeypatch.context() as p:
+        p.setattr(mp, "get_context", lambda _: utils.DummyContext(wid=0))
+        s = simplifier.Simp(crash_dir=Path("/"), output_dir=Path("/"), target=target)
+        p.setattr(s, "_workers", workers)
+        assert not cast(utils.DummyQueue[simplifier.Metrics], s._result_queue).canceled
+        assert all(
+            not w[0].terminated and not w[0].joined and w[0].timeout is None and not w[1].canceled
+            for w in workers
+        )
+        s.terminate_workers()
+        assert cast(utils.DummyQueue[simplifier.Metrics], s._result_queue).canceled
+        assert all(
+            w[0].terminated and w[0].joined and w[0].timeout == 1 and w[1].canceled for w in workers
+        )
+
+        previous_workers = copy.copy(workers)
+        s.terminate_workers()
+        assert workers == previous_workers
+
+
+def test_simplify_inner_loop(monkeypatch: pytest.MonkeyPatch) -> None:
+    def target(_: bytes) -> None:
+        pass  # pragma: no cover
+
+    def run_target(_target: Callable[[bytes], None], data: bytes) -> Optional[simplifier.Metrics]:
+        return simplifier.Metrics(data, set())
+
+    args: ArgsType = (
+        b"deadbeef",
+        utils.DummyQueue(),
+        utils.DummyQueue(),
+    )
+
+    workers: list[tuple[utils.DummyProcess[ArgsType], utils.DummyQueue[simplifier.Metrics]]] = [
+        (utils.DummyProcess(target=target, args=args), utils.DummyQueue()),
+        (utils.DummyProcess(target=target, args=args), utils.DummyQueue(length=1)),
+    ]
+
+    value_1 = b"aaaaaa"
+    value_2 = b"aaaaa"
+
+    result_queue: utils.DummyQueue[simplifier.Metrics] = utils.DummyQueue()
+    result_queue.put(simplifier.Metrics(value_1, coverage=set()))
+    result_queue.put(simplifier.Metrics(value_2, coverage=set()))
+
+    # Make second worker queue full
+    workers[1][1].put(simplifier.Metrics(data=b"unrelated", coverage=set()))
+
+    with monkeypatch.context() as p:
+        p.setattr(simplifier, "run_target", run_target)
+        s = simplifier.Simp(crash_dir=Path("/"), output_dir=Path("/"), target=target, max_time=1)
+        p.setattr(s, "_result_queue", result_queue)
+        p.setattr(s, "_workers", workers)
+
+        assert s._simplify(value_1) == value_2
+        assert not workers[0][1].empty()
+        assert workers[0][1].get().data == value_2