+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Crash result simplification (#30)
 - Configurable simplification steps (#54)
+- Parallel simplification (#53)
 
 ### Changed